decimals work

+++ models/beer.js
@@ -2,7 +2,7 @@ module.exports = function(sequelize, DataTypes) {
   var Beer = sequelize.define("Beer", {
     beer_name: DataTypes.STRING,
     brewery: DataTypes.STRING,
-    abv: DataTypes.INTEGER,
+    abv: DataTypes.FLOAT,
     user_id: DataTypes.STRING
   });
   return Beer;
